+++ src/main/java/io/mbab/sda/groupproject2/Application.java
@@ -1,13 +1,11 @@
-package io.mbab.sda.groupproject;
+package io.mbab.sda.groupproject2;
 
-import io.mbab.sda.groupproject.config.Configuration;
-import io.mbab.sda.groupproject.config.JpaUtil;
-import io.mbab.sda.groupproject.menu.CustomScanner;
-import io.mbab.sda.groupproject.menu.action.MainAction;
-import io.mbab.sda.groupproject.menu.MenuActionContext;
-import io.mbab.sda.groupproject.repository.CrudRepositoryFactory;
-
-import java.util.Scanner;
+import io.mbab.sda.groupproject2.config.Configuration;
+import io.mbab.sda.groupproject2.config.JpaUtil;
+import io.mbab.sda.groupproject2.menu.CustomScanner;
+import io.mbab.sda.groupproject2.menu.action.MainAction;
+import io.mbab.sda.groupproject2.menu.MenuActionContext;
+import io.mbab.sda.groupproject2.repository.CrudRepositoryFactory;
 
 public class Application {
