+++ src/main/java/io/mbab/sda/groupproject2/config/Configuration.java
@@ -1,7 +1,7 @@
-package io.mbab.sda.groupproject.config;
+package io.mbab.sda.groupproject2.config;
 
 import com.mysql.cj.jdbc.MysqlDataSource;
-import io.mbab.sda.groupproject.entity.City;
+import io.mbab.sda.groupproject2.entity.Album;
 
 import javax.sql.DataSource;
 
@@ -9,7 +9,7 @@ public final class Configuration {
   private static final String DB_URL =
       "jdbc:mysql://localhost:3306/group_project?useSSL=false&serverTimezone=UTC&characterEncoding=UTF-8&allowPublicKeyRetrieval=true";
   private static final String DB_USER_NAME = "root";
-  private static final String DB_PASSWORD = "root";
+  private static final String DB_PASSWORD = "N90de5N90de5";
 
   private static DataSource dataSource;
 
@@ -28,6 +28,6 @@ public final class Configuration {
 
   /* W tej metodzie należy zwrócić liste klas które są encjami, aplikacja samodzielnie ich nie zmapuje */
   public static Class[] getEntityClass() {
-    return new Class[] {City.class};
+    return new Class[] {Album.class};
   }
 }
