+++ src/main/java/io/mbab/sda/groupproject2/menu/action/MainAction.java
@@ -1,7 +1,7 @@
-package io.mbab.sda.groupproject.menu.action;
+package io.mbab.sda.groupproject2.menu.action;
 
-import io.mbab.sda.groupproject.menu.CustomScanner;
-import io.mbab.sda.groupproject.menu.MenuActionContext;
+import io.mbab.sda.groupproject2.menu.CustomScanner;
+import io.mbab.sda.groupproject2.menu.MenuActionContext;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
@@ -13,8 +13,8 @@ public class MainAction implements MenuAction {
   @Override
   public void execute() {
     System.out.println("0) Zamknij aplikację");
-    System.out.println("1) Dodaj miasto");
-    System.out.println("2) Wyswietl miasta");
+    System.out.println("1) Dodaj Album");
+    System.out.println("2) Wyswietl nazwe Albumu");
 
     var input = scanner.nextLine();
 
@@ -24,7 +24,7 @@ public class MainAction implements MenuAction {
     }
 
     if (input.equals("1")) {
-      ctx.use(CreateCityAction.class).execute();
+      ctx.use(CreateAlbumAction.class).execute();
       return;
     }
